% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_stationary.R
\docType{data}
\name{data_stationary}
\alias{data_stationary}
\title{test sample data}
\format{
A data frame with 1000 rows and 4 variables:
\describe{
  \item{Date}{A simulated data}
  \item{y}{outcome}
  \item{c}{covarites}
  \item{x}{treatment}
}
}
\usage{
data_stationary
}
\description{
test sample data
}
\keyword{datasets}
