% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_closepoints.R
\name{merge_closepoints}
\alias{merge_closepoints}
\title{merge close points}
\usage{
merge_closepoints(points, band)
}
\arguments{
\item{points}{points to be merge}

\item{band}{set up a band for merge}
}
\description{
merge close points
}
