% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printer.R
\name{printer}
\alias{printer}
\title{test printer}
\usage{
printer(r, x)
}
\arguments{
\item{r}{Variable r}

\item{x}{Variable x}
}
\value{
A tibble
}
\description{
test printer
}
\examples{
printer(x = rnorm(5), r = rnorm(5))
}
