% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.SSM_unanimous_cpts.R
\name{run.SSM_unanimous_cpts}
\alias{run.SSM_unanimous_cpts}
\title{SSM function}
\usage{
run.SSM_unanimous_cpts(
  data_ss,
  formula_var,
  ss_param_temp,
  max_iteration = 100,
  cpt_learning_param = list(cpt_method = "mean", burnin = 1/10, mergeband = 20,
    convergence_cri = 10),
  dlm_option = "smooth",
  printFlag = T
)
}
\arguments{
\item{data_ss}{contains all information, and only selected variables in formula_var enters the statespace model}

\item{formula_var}{select variables from <data_ss> into the statespace model}

\item{ss_param_temp}{<m0>: initial values for states}

\item{max_iteration}{control for the convergence of changepoints, a positive integer}

\item{cpt_learning_param}{<cpt_learning_param>:  <cpt_method> either "mean" or "meanvar"}

\item{dlm_option}{...}

\item{printFlag}{T / F}
}
\value{

}
\description{
SSM function
}
