% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.SSMimpute_unanimous_cpts.R
\name{run.SSMimpute_unanimous_cpts}
\alias{run.SSMimpute_unanimous_cpts}
\title{SSMimpute: state space model direct imputation on missing data in covariates}
\usage{
run.SSMimpute_unanimous_cpts(
  data_ss_ori,
  formula_var,
  ss_param_temp,
  initial_imputation_option = "StructTS",
  estimate_convergence_cri = 0.01,
  lik_convergence_cri = 0.01,
  stepsize_for_newpart = 1/3,
  max_iteration = 100,
  cpt_learning_param = list(cpt_method = "mean", burnin = 1/10, mergeband = 20,
    convergence_cri = 15),
  cpt_initial_guess_option = "ignore",
  dlm_option = "smooth",
  m = 5,
  seed = 1,
  printFlag = T
)
}
\arguments{
\item{data_ss_ori}{contains all information, and only selected variables in formula_var enters the statespace model}

\item{formula_var}{select variables from <data_ss_ori> into the state space model}

\item{ss_param_temp}{A list of parameters, details below}

\item{initial_imputation_option}{for the first iteration of imputing missing y, choose StructTS or others, and can't be "ignore"}

\item{estimate_convergence_cri}{critic value for convergence check, default 0.01}

\item{lik_convergence_cri}{critic value for convergence check, default 0.01}

\item{stepsize_for_newpart}{stepsize specified, default 1/3}

\item{max_iteration}{max iteration, default 100}

\item{cpt_learning_param}{<cpt_method> either "mean" or "meanvar"}

\item{cpt_initial_guess_option}{option for initially learning cpts in preparation period}

\item{dlm_option}{choose between smooth or filter}

\item{m}{number of draws for multiple imputation}

\item{seed}{random seed}

\item{printFlag}{whether we need to print the Flag plots.}
}
\value{
A list
}
\description{
SSMimpute: state space model direct imputation on missing data in covariates
}
\details{
<m0> initial values for states <C0>: initial values for variance of states <inits>: initial values for the estimating of all NA terms, via maximizing likelihood
}
