% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Test sample data}
\format{
A data frame with 999 rows and 8 variables:
\describe{
  \item{x}{the index}
  \item{y}{outcome}
  ...
}
}
\source{
\url{http://www.xxx.com/}
}
\usage{
sample_data
}
\description{
A sample data set for imputation test
}
\keyword{datasets}
