% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_space_SSMimpute.R
\docType{data}
\name{data_space_SSMimpute}
\alias{data_space_SSMimpute}
\title{test sample data}
\format{
A example data frame
\describe{
  \item{x}{x}
  \item{y}{x}
  ...
}
}
\source{
\url{http://www.xxx.com/}
}
\usage{
data_space_SSMimpute
}
\description{
test sample data
}
\keyword{datasets}
