% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_space_SSMimpute.R
\docType{data}
\name{data_space_SSMimpute}
\alias{data_space_SSMimpute}
\title{test sample data}
\format{
A data frame with 999 rows and 8 variables:
\describe{
  \item{X}{index}
  \item{y}{outcome}
  \item{c}{covarites}
  \item{c_1}{c value at X-1}
  \item{x}{x value}
  \item{x_1}{x value at X-1}
  \item{y_1}{y value at X-1}
  \item{Date}{Date of real world}
  ...
}
}
\source{
\url{http://www.xxx.com/}
}
\usage{
data_space_SSMimpute
}
\description{
test sample data
}
\keyword{datasets}
