% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{test sample data}
\format{
A data frame with 100 rows and 2 variables:
\describe{
  \item{x}{x}
  \item{y}{x}
  ...
}
}
\source{
\url{http://www.xxx.com/}
}
\usage{
sample_data
}
\description{
A sample data set for imputation
}
\keyword{datasets}
